#22: update noted noted to have an id and title
