+++ lib/noted_models.dart
@@ -3,5 +3,5 @@ library noted_models;
 export 'auth/auth.dart';
 export 'rich_text/rich_text.dart';
 export 'settings/settings.dart';
-export 'spaces/spaces.dart';
+export 'notes/notes.dart';
 export 'theme/theme.dart';
