+++ lib/notes/noted_note.mapper.dart
@@ -3,7 +3,67 @@
 // ignore_for_file: type=lint
 // ignore_for_file: unused_element
 
-part of 'notebook_note.dart';
+part of 'noted_note.dart';
+
+class NotedNoteMapper extends ClassMapperBase<NotedNote> {
+  NotedNoteMapper._();
+
+  static NotedNoteMapper? _instance;
+  static NotedNoteMapper ensureInitialized() {
+    if (_instance == null) {
+      MapperContainer.globals.use(_instance = NotedNoteMapper._());
+      NotebookNoteMapper.ensureInitialized();
+    }
+    return _instance!;
+  }
+
+  static T _guard<T>(T Function(MapperContainer) fn) {
+    ensureInitialized();
+    return fn(MapperContainer.globals);
+  }
+
+  @override
+  final String id = 'NotedNote';
+
+  static String _$id(NotedNote v) => v.id;
+  static const Field<NotedNote, String> _f$id = Field('id', _$id);
+  static String _$title(NotedNote v) => v.title;
+  static const Field<NotedNote, String> _f$title = Field('title', _$title);
+
+  @override
+  final Map<Symbol, Field<NotedNote, dynamic>> fields = const {
+    #id: _f$id,
+    #title: _f$title,
+  };
+
+  static NotedNote _instantiate(DecodingData data) {
+    throw MapperException.missingSubclass(
+        'NotedNote', 'plugin', '${data.value['plugin']}');
+  }
+
+  @override
+  final Function instantiate = _instantiate;
+
+  static NotedNote fromMap(Map<String, dynamic> map) {
+    return _guard((c) => c.fromMap<NotedNote>(map));
+  }
+
+  static NotedNote fromJson(String json) {
+    return _guard((c) => c.fromJson<NotedNote>(json));
+  }
+}
+
+mixin NotedNoteMappable {
+  String toJson();
+  Map<String, dynamic> toMap();
+  NotedNoteCopyWith<NotedNote, NotedNote, NotedNote> get copyWith;
+}
+
+abstract class NotedNoteCopyWith<$R, $In extends NotedNote, $Out>
+    implements ClassCopyWith<$R, $In, $Out> {
+  $R call({String? id, String? title});
+  NotedNoteCopyWith<$R2, $In, $Out2> $chain<$R2, $Out2>(Then<$Out2, $R2> t);
+}
 
 class NotebookNoteMapper extends SubClassMapperBase<NotebookNote> {
   NotebookNoteMapper._();
@@ -41,7 +101,7 @@ class NotebookNoteMapper extends SubClassMapperBase<NotebookNote> {
   };
 
   @override
-  final String discriminatorKey = 'space';
+  final String discriminatorKey = 'plugin';
   @override
   final dynamic discriminatorValue = 'notebook';
   @override
