+++ lib/notes/notes.dart
@@ -1,0 +1,1 @@
+export 'noted_note.dart';
