--- lib/spaces/notebook/notebook.dart
@@ -1,1 +1,0 @@
-export 'notebook_note.dart';
