--- lib/spaces/noted_note.dart
@@ -1,8 +1,0 @@
-import 'package:dart_mappable/dart_mappable.dart';
-
-part 'noted_note.mapper.dart';
-
-@MappableClass(discriminatorKey: 'space')
-abstract class NotedNote with NotedNoteMappable {
-  const NotedNote();
-}
